Formatted Essay Tree can be saved to database
My current updates have made it so that I can retrieve the formatted essay tree off of jquery, post it to a php page, and save it to a database from there. Also, the essay names will show up as a hyperlink on the account page, giving the user access go back to their saved essay. Currently, work now needs to be done to get the essay tree off of the database, and back into a tree structure.